+++ examples/all-possible-containers/src/components/FunctionComponent.js
@@ -1,0 +1,13 @@
+import React from 'react'
+import { EDIT_ME } from './_editMe'
+
+const FunctionComponent = () => (
+  <div>
+    <fieldset>
+      <legend>Function Component</legend>
+      {EDIT_ME}
+    </fieldset>
+  </div>
+)
+
+export default FunctionComponent