+++ examples/all-possible-containers/src/components/PureClassComponent.js
@@ -1,0 +1,17 @@
+import React from 'react'
+import { EDIT_ME } from './_editMe'
+
+class ClassComponent extends React.PureComponent {
+  render() {
+    return (
+      <div>
+        <fieldset>
+          <legend>Pure Class Component</legend>
+          {EDIT_ME}
+        </fieldset>
+      </div>
+    )
+  }
+}
+
+export default ClassComponent