+++ examples/all-possible-containers/src/components/_editMe.js
@@ -1,0 +1,1 @@
+export const EDIT_ME = 'Edit me'