+++ src/configuration.js
@@ -1,5 +1,6 @@
 const configuration = {
   logLevel: 'error',
+  pureSFC: false,
 }
 
 export default configuration