+++ src/proxy/createClassProxy.js
@@ -16,6 +16,7 @@ import {
   proxyClassCreator,
 } from './utils'
 import { inject, checkLifeCycleMethods, mergeComponents } from './inject'
+import config from '../configuration'
 
 const has = Object.prototype.hasOwnProperty
 
@@ -108,6 +109,8 @@ function createClassProxy(InitialComponent, proxyKey, options) {
   let savedDescriptors = {}
   let injectedMembers = {}
   let proxyGeneration = 0
+  let classUpdatePostponed = null
+  let instancesCount = 0
   let isFunctionalComponent = !isReactClass(InitialComponent)
 
   let lastInstance = null
@@ -115,11 +118,16 @@ function createClassProxy(InitialComponent, proxyKey, options) {
   function postConstructionAction() {
     this[GENERATION] = 0
 
+    lastInstance = this
+    // is there is an update pending
+    if (classUpdatePostponed) {
+      const callUpdate = classUpdatePostponed
+      classUpdatePostponed = null
+      callUpdate()
+    }
     // As long we can't override constructor
     // every class shall evolve from a base class
     inject(this, proxyGeneration, injectedMembers)
-
-    lastInstance = this
   }
 
   function proxiedUpdate() {
@@ -162,6 +170,7 @@ function createClassProxy(InitialComponent, proxyKey, options) {
     'componentDidMount',
     target => {
       target[PROXY_IS_MOUNTED] = true
+      instancesCount++
     },
   )
   const componentDidUpdate = lifeCycleWrapperFactory(
@@ -172,6 +181,7 @@ function createClassProxy(InitialComponent, proxyKey, options) {
     'componentWillUnmount',
     target => {
       target[PROXY_IS_MOUNTED] = false
+      instancesCount--
     },
   )
 
@@ -189,7 +199,7 @@ function createClassProxy(InitialComponent, proxyKey, options) {
     } else if (isFunctionalComponent) {
       result = CurrentComponent(this.props, this.context)
     } else {
-      result = CurrentComponent.prototype.render.call(this)
+      result = (CurrentComponent.prototype.render || this.render).call(this)
     }
 
     return renderOptions.componentDidRender.call(this, result)
@@ -230,12 +240,14 @@ function createClassProxy(InitialComponent, proxyKey, options) {
       const result = CurrentComponent(props, context)
 
       // simple SFC
-      if (!CurrentComponent.contextTypes) {
-        if (!ProxyFacade.isStatelessFunctionalProxy) {
-          setSFPFlag(ProxyFacade, true)
-        }
+      if (config.pureSFC) {
+        if (!CurrentComponent.contextTypes) {
+          if (!ProxyFacade.isStatelessFunctionalProxy) {
+            setSFPFlag(ProxyFacade, true)
+          }
 
-        return renderOptions.componentDidRender(result)
+          return renderOptions.componentDidRender(result)
+        }
       }
       setSFPFlag(ProxyFacade, false)
 
@@ -308,9 +320,10 @@ function createClassProxy(InitialComponent, proxyKey, options) {
       return
     }
 
+    isFunctionalComponent = !isReactClass(NextComponent)
+
     proxies.set(NextComponent, proxy)
 
-    isFunctionalComponent = !isReactClass(NextComponent)
     proxyGeneration++
 
     // Save the next constructor so we call it
@@ -343,17 +356,26 @@ function createClassProxy(InitialComponent, proxyKey, options) {
     if (isFunctionalComponent || !ProxyComponent) {
       // nothing
     } else {
-      checkLifeCycleMethods(ProxyComponent, NextComponent)
-      Object.setPrototypeOf(ProxyComponent.prototype, NextComponent.prototype)
-      defineProxyMethods(ProxyComponent, NextComponent.prototype)
-      if (proxyGeneration > 1) {
-        injectedMembers = mergeComponents(
-          ProxyComponent,
-          NextComponent,
-          InitialComponent,
-          lastInstance,
-          injectedMembers,
-        )
+      const classHotReplacement = () => {
+        checkLifeCycleMethods(ProxyComponent, NextComponent)
+        Object.setPrototypeOf(ProxyComponent.prototype, NextComponent.prototype)
+        defineProxyMethods(ProxyComponent, NextComponent.prototype)
+        if (proxyGeneration > 1) {
+          injectedMembers = mergeComponents(
+            ProxyComponent,
+            NextComponent,
+            InitialComponent,
+            lastInstance,
+            injectedMembers,
+          )
+        }
+      }
+
+      // Was constructed once
+      if (instancesCount > 0) {
+        classHotReplacement()
+      } else {
+        classUpdatePostponed = classHotReplacement
       }
     }
   }